Create a new Chart.Ticks namespace. Add formatters to Chart.Ticks.formatters.
