+++ src/core/core.scale.js
@@ -47,7 +47,7 @@ module.exports = function(Chart) {
 			labelOffset: 0,
 			// We pass through arrays to be rendered as multiline labels, we convert Others to strings here.
 			callback: function(value) {
-				return helpers.isArray(value) ? value : '' + value;
+				return Chart.Ticks.formatters.values(value);
 			}
 		}
 	};
